Fix get buffer time return type

+++ src/ofxNatNet.h
@@ -116,7 +116,7 @@ public:
 	}
 
 	void setBufferTime(float sec);
-	int getBufferTime();
+	float getBufferTime();
 	
 	void setTimeout(float timeout);
